+++ src/dune_engine/dune_project.ml
@@ -165,6 +165,8 @@ let equal = ( == )
 
 let hash = Hashtbl.hash
 
+let parsing_context t = t.parsing_context
+
 let packages t = t.packages
 
 let version t = t.version