+++ test/blackbox-tests/test-cases/ctypes/exe-pkg_config-multiple-fd.t/example.ml
@@ -1,0 +1,4 @@
+let () =
+  let r1 = C.Functions_sequential.add2 0 in
+  let r2 = C.Functions_unlocked.add4 r1 in
+  Printf.printf "%d\n" r2