+++ test/blackbox-tests/test-cases/ctypes/exe-pkg_config.t/libexample/example.h
@@ -1,0 +1,1 @@
+int example_add2(int);