+++ test/blackbox-tests/test-cases/ctypes/exe-vendored-override-types-generated.t/run.t
@@ -1,0 +1,13 @@
+Generate cstubs for a "vendored" library.
+
+We have a dummy C library hosted entirely in the 'vendor' directory and use
+the ctypes instrumentation and description language to generate bindings for
+it.
+
+This is the version that builds into an executable.
+
+This test is identical to exe-vendored.t except it overrides the
+generated_types stanza field.
+
+  $ dune exec ./example.exe
+  4